+++ common/src/main/java/net/SpectrumFATM/black_archive/blockentity/ModModels.java
@@ -4,8 +4,10 @@ import dev.architectury.injectables.annotations.ExpectPlatform;
 import net.SpectrumFATM.BlackArchive;
 import net.SpectrumFATM.black_archive.blockentity.door.PillarDoorModel;
 import net.SpectrumFATM.black_archive.blockentity.door.RaniDoorModel;
+import net.SpectrumFATM.black_archive.blockentity.door.SIDRATDoorModel;
 import net.SpectrumFATM.black_archive.blockentity.shell.PillarShellModel;
 import net.SpectrumFATM.black_archive.blockentity.shell.RaniShellModel;
+import net.SpectrumFATM.black_archive.blockentity.shell.SIDRATModel;
 import net.SpectrumFATM.black_archive.blockentity.shell.ShellEntryRegistry;
 import net.minecraft.client.model.TexturedModelData;
 import net.minecraft.client.render.entity.model.EntityModelLayer;
@@ -21,12 +23,14 @@ public class ModModels {
 
     public static PillarShellModel PILLAR_EXT_MODEL;
     public static RaniShellModel RANI_EXT_MODEL;
+    public static SIDRATModel SIDRAT_EXT_MODEL;
 
     public static PillarDoorModel PILLAR_INT_MODEL;
     public static RaniDoorModel RANI_INT_MODEL;
+    public static SIDRATDoorModel SIDRAT_INT_MODEL;
 
-    public static EntityModelLayer PILLAR_EXT, RANI_EXT;
-    public static EntityModelLayer PILLAR_INT, RANI_INT;
+    public static EntityModelLayer PILLAR_EXT, RANI_EXT, SIDRAT_EXT;
+    public static EntityModelLayer PILLAR_INT, RANI_INT, SIDRAT_INT;
 
     public static void init() {
         if (Platform.isClient()) {
@@ -37,6 +41,9 @@ public class ModModels {
         PILLAR_INT = register(new EntityModelLayer(new Identifier(BlackArchive.MOD_ID, "pillar_int"), "pillar_int"), PillarDoorModel::getTexturedModelData);
         RANI_EXT = register(new EntityModelLayer(new Identifier(BlackArchive.MOD_ID, "rani_ext"), "rani_ext"), RaniShellModel::getTexturedModelData);
         RANI_INT = register(new EntityModelLayer(new Identifier(BlackArchive.MOD_ID, "rani_int"), "rani_int"), RaniDoorModel::getTexturedModelData);
+        SIDRAT_INT = register(new EntityModelLayer(new Identifier(BlackArchive.MOD_ID, "sidrat_int"), "sidrat_int"), SIDRATDoorModel::getTexturedModelData);
+        SIDRAT_EXT = register(new EntityModelLayer(new Identifier(BlackArchive.MOD_ID, "sidrat_ext"), "sidrat_ext"), SIDRATModel::getTexturedModelData);
+
     }
 
     public static void setupModelInstances(EntityModelLoader entityModels) {
@@ -44,6 +51,8 @@ public class ModModels {
         ModModels.PILLAR_INT_MODEL = new PillarDoorModel(entityModels.getModelPart(ModModels.PILLAR_INT));
         ModModels.RANI_EXT_MODEL = new RaniShellModel(entityModels.getModelPart(ModModels.RANI_EXT));
         ModModels.RANI_INT_MODEL = new RaniDoorModel(entityModels.getModelPart(ModModels.RANI_INT));
+        ModModels.SIDRAT_EXT_MODEL = new SIDRATModel(entityModels.getModelPart(ModModels.SIDRAT_EXT));
+        ModModels.SIDRAT_INT_MODEL = new SIDRATDoorModel(entityModels.getModelPart(ModModels.SIDRAT_INT));
         ShellEntryRegistry.init();
     }